Move new inline link under same folder
